pyproject.toml
[tool.poetry]
name = "finance-site"
version = "0.1.0"
description = ""
authors = ["muratbek2255 <muratbek2255@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.3"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
Pillow = "^9.3.0"
pyotp = "^2.7.0"
django-environ = "^0.9.0"
twilio = "^7.15.3"
django-filter = "^22.1"
celery = "^5.2.7"
drf-yasg = "^1.21.4"
gunicorn = "^20.1.0"
django-cors-headers = "^3.13.0"
redis = "^4.3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
